Separate into TypeChecker/Utils, only include exposed definitions in final map
